+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add a product flavor for TamilNadu app
 - Bump google play services location to v21.0.1
 - Bump Sentry Gradle Plugin to v3.2.1
+- Move splash screen lottie config to XML
 
 ## 2022-10-31-8480